fix: clean up diarization test infrastructure
- Extract duplicate fixture helpers (makeFixtureSamples, writeFixtureAudio,
  chunk, fixtureAudioFileURL) into shared DiarizationTestFixtures enum
- Add LSEENDDiarizer.hasActiveSession property to replace fragile
  Mirror-based reflection for checking private _session state
- Use fixed seed (42) for srand48 in SortformerTests to make
  testFeatureProviderEquivalency failures reproducible
- Add XCTSkipIf guards to Sortformer enrollment tests to match
  LS-EEND pattern (was XCTAssertNotNil which hard-fails when fixture
  audio doesn't trigger speaker detection on a given host)
